Handle styling from diagram
Each item type should bring its own style node implementation (wrapper).
